Move util tests to separate test module
